+++ src/cmd/login/login.rs
@@ -1,13 +1,13 @@
 use crate::cmd::{
     args::{Cli, Commands},
-    config::{Configuration, RepositoryConfig},
+    config::{Configuration, RepositoryConfig, VcsConfig},
 };
 use eyre::{eyre, Context, Result};
 use gr_bin::{
     git::git::LocalRepository,
     vcs::common::{init_vcs, VersionControlSettings},
 };
-use std::{io, thread::sleep, time::Duration};
+use std::{collections::HashMap, io, io::Write, thread::sleep, time::Duration};
 use tracing::instrument;
 
 #[instrument(skip_all, fields(command = ?args.command))]
@@ -15,12 +15,16 @@ pub fn login(args: Cli, mut conf: Configuration) -> Result<()> {
     let Cli { command, dir, .. } = args;
     if let Commands::Login {
         hostname,
+        vcs_type,
         repo: repo_name,
         token,
     } = command
     {
         // Get hostname and repo and initialize VCS
-        let settings = VersionControlSettings::default();
+        let settings = VersionControlSettings {
+            vcs_type: vcs_type.clone(),
+            ..VersionControlSettings::default()
+        };
         let (hostname, repo) = if let Some(hostname) = hostname {
             (hostname, repo_name.clone().unwrap_or_default())
         } else {
@@ -29,7 +33,7 @@ pub fn login(args: Cli, mut conf: Configuration) -> Result<()> {
 
             (hostname, repo)
         };
-        let vcs = init_vcs(hostname.clone(), repo, settings);
+        let vcs = init_vcs(hostname.clone(), repo, settings)?;
 
         // If the token arg is passed, validate and use that
         let token = if let Some(token) = token {
@@ -41,7 +45,8 @@ pub fn login(args: Cli, mut conf: Configuration) -> Result<()> {
                 "To login to {}, create a token and copy the token value.",
                 hostname
             );
-            if !url.contains("scopes=") {
+            // TODO: this is a dirty hack
+            if hostname == "bitbucket.org" {
                 println!("The token needs account, workspace and project read, pull request read and write permissions.");
                 println!("You have to enter you username and the token separated with a colon (e.g. user:ATBB...).");
             }
@@ -53,13 +58,17 @@ pub fn login(args: Cli, mut conf: Configuration) -> Result<()> {
             sleep(Duration::from_millis(500));
 
             // Read the token from the user
-            let mut token = String::new();
+            let mut token;
+            let mut stdout = io::stdout();
             let stdin = io::stdin();
             loop {
-                print!("Paste the token here: ");
+                write!(stdout, "Paste the token here: ")?;
+                stdout.flush()?;
+                token = String::new();
                 stdin
                     .read_line(&mut token)
                     .wrap_err("Reading the token failed.")?;
+                token = token.trim().to_string();
                 match vcs.validate_token(&token) {
                     Ok(_) => break,
                     Err(err) => println!("{}", err),
@@ -71,20 +80,30 @@ pub fn login(args: Cli, mut conf: Configuration) -> Result<()> {
         // Modify the token in the configuration
         let host_conf = conf.vcs.entry(hostname);
         match repo_name {
-            Some(repo) => host_conf.and_modify(|h| {
-                h.repositories
-                    .entry(repo)
-                    .and_modify(|r| {
-                        r.auth = Some(token.clone());
+            Some(repo) => {
+                host_conf.and_modify(|h| {
+                    h.repositories
+                        .entry(repo)
+                        .and_modify(|r| {
+                            r.auth = Some(token.clone());
+                        })
+                        .or_insert(RepositoryConfig {
+                            auth: Some(token),
+                            default_branch: None,
+                        });
+                });
+            }
+            None => {
+                host_conf
+                    .and_modify(|h| {
+                        h.auth = token.clone();
                     })
-                    .or_insert(RepositoryConfig {
-                        auth: Some(token),
-                        default_branch: None,
+                    .or_insert(VcsConfig {
+                        auth: token,
+                        repositories: HashMap::default(),
+                        vcs_type,
                     });
-            }),
-            None => host_conf.and_modify(|h| {
-                h.auth = token;
-            }),
+            }
         };
         conf.save()?;
         println!("Authentication token saved.");